+++ packages/hedera-wallet-snap/packages/snap/src/services/hedera.ts
@@ -35,11 +35,12 @@ import { AccountInfo } from '../types/account';
 import { ApproveAllowanceAssetDetail } from '../types/params';
 
 export type SimpleTransfer = {
-  // HBAR or Token ID (as string)
-  asset: string;
+  assetType: 'HBAR' | 'TOKEN' | 'NFT';
   to: string;
   // amount must be in low denom
   amount: number;
+  // Token or NFT ID (as string)
+  assetId?: string;
 };
 
 export type Token = {